push

+++ src/setup.sh
@@ -8,10 +8,12 @@ echo "Installing the necessary packages"
 sudo apt install -y curl dos2unix wget git dpkg-dev &> /tmp/apt.txt
 sudo apt clean &> /tmp/apt.txt
 #
-echo "Removing unnecessary packages"
-apt purge --remove *dotnet* -y &> /tmp/apt.txt
-sudo rm -rf /usr/share/dotnet &> /tmp/apt.txt
-sudo rm -rf /usr/local/lib/android &> /tmp/apt.txt
+if [ $INPUT_SPACE == 'true' ];then
+    echo "Removing unnecessary packages"
+    apt purge --remove *dotnet* -y &> /tmp/apt.txt
+    sudo rm -rf /usr/share/dotnet &> /tmp/apt.txt
+    sudo rm -rf /usr/local/lib/android &> /tmp/apt.txt
+fi
 # Autoremove
 sudo apt-get -qq autoremove --purge &> /tmp/apt.txt
 echo "removing the swap"
